refactor(clash): split clash_solid.rs's geometric helpers into clash_solid_geom.rs
The projected trust-band fix pushed clash_solid.rs to 445 lines, past the
module_size_ratchet's 400-line limit with no allowlist row. Move
tri_volume, component_groups, and operand_near_band (413 lines together,
mostly their doc comments) into a sibling clash_solid_geom.rs, leaving
intersection_solid and the public API types in clash_solid.rs.

clash_solid.rs: 445 -> 288 lines. clash_solid_geom.rs: 173 lines. Both
under the ratchet's 400-line limit, so no allowlist row is needed.
